requirements.txt
django-cors-headers==4.3.1
azure-cognitiveservices-speech==1.36.0
python-dotenv==1.0.0
openai==1.14.2
azure-functions==1.18.0
psycopg2-binary==2.9.9
django==4.1.7
requests==2.31.0
uuid==1.30
